+++ src/main/java/com/prgrms/artzip/review/service/ReviewService.java
@@ -10,8 +10,8 @@ import com.prgrms.artzip.review.domain.Review;
 import com.prgrms.artzip.review.domain.ReviewPhoto;
 import com.prgrms.artzip.review.domain.repository.ReviewPhotoRepository;
 import com.prgrms.artzip.review.domain.repository.ReviewRepository;
-import com.prgrms.artzip.review.dto.ReviewCreateRequest;
-import com.prgrms.artzip.review.dto.ReviewCreateResponse;
+import com.prgrms.artzip.review.dto.request.ReviewCreateRequest;
+import com.prgrms.artzip.review.dto.response.ReviewCreateResponse;
 import com.prgrms.artzip.user.domain.User;
 import com.prgrms.artzip.user.domain.repository.UserRepository;
 import java.io.IOException;
